+++ src/main/java/com/ayakacraft/carpetayakaaddition/commands/address/Address.java
@@ -29,16 +29,21 @@ import net.minecraft.world.World;
 //TODO update in 1.15.2
 public class Address extends AbstractAddress<RegistryKey<World>> {
 
-    public Address(String id, String dim, Vec3d pos, String desc, long weight) {
+    public Address(String id, String dim, Vec3d pos, String desc, int weight) {
         super(id, dim, pos, desc, weight);
     }
 
-    public Address(String id, RegistryKey<World> dim, Vec3d pos, String desc, long weight) {
-        super(id, dim.getValue().toString(), pos, desc, weight);
+    public Address(String id, RegistryKey<World> dim, Vec3d pos, String desc, int weight) {
+        super(id, dim, pos, desc, weight);
     }
 
     public Address(AddressOld old) {
         super(old);
+    }
+
+    @Override
+    protected String transDim(RegistryKey<World> worldRegistryKey) {
+        return worldRegistryKey.getValue().toString();
     }
 
     @Override