+++ src/main/java/com/ayakacraft/carpetayakaaddition/utils/InitializedPerTick.java
@@ -25,32 +25,10 @@ import com.ayakacraft.carpetayakaaddition.CarpetAyakaServer;
 @FunctionalInterface
 public interface InitializedPerTick {
 
+    static TickTask.RunPerTickTask createTickTask(CarpetAyakaServer modServer, InitializedPerTick perTick) {
+        return new TickTask.RunPerTickTask(modServer, perTick::init, true);
+    }
+
     void init();
-
-    default TickTask getInitTask(CarpetAyakaServer modServer) {
-        return new CounterInitTask(modServer, this);
-    }
-
-    class CounterInitTask extends TickTask {
-
-        private final InitializedPerTick initializedPerTick;
-
-        public CounterInitTask(CarpetAyakaServer modServer, InitializedPerTick initializedPerTick) {
-            super(modServer);
-            this.initializedPerTick = initializedPerTick;
-        }
-
-        @Override
-        public void cancel() {
-            super.cancel();
-            initializedPerTick.init();
-        }
-
-        @Override
-        public void tick() {
-            initializedPerTick.init();
-        }
-
-    }
 
 }