+++ src/main/java/com/ayakacraft/carpetayakaaddition/utils/text/TextUtils.java
@@ -24,7 +24,10 @@ import com.ayakacraft.carpetayakaaddition.utils.preprocess.PreprocessPattern;
 import com.ayakacraft.carpetayakaaddition.utils.translation.Translator;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.text.*;
+import net.minecraft.text.ClickEvent;
+import net.minecraft.text.HoverEvent;
+import net.minecraft.text.MutableText;
+import net.minecraft.text.Text;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -118,10 +121,9 @@ public final class TextUtils {
         sendMessageToServer(server, textForServer);
         server.getPlayerManager().getPlayerList().forEach(player -> {
             Text t = textFunction.apply(player);
-            if (t == null) {
-                return;
+            if (t != null) {
+                player.sendMessage(t, overlay);
             }
-            player.sendMessage(t, overlay);
         });
     }
 
@@ -134,10 +136,12 @@ public final class TextUtils {
         );
     }
 
-    public static Style runCommand(Style style, String command) {
-        return style
-                .withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, command))
-                .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Text.literal(command)));
+    public static Text withCommand(MutableText text, String command) {
+        return text.styled(style ->
+                style
+                        .withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, command))
+                        .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Text.literal(command)))
+        );
     }
 
 }