+++ versions/1.15.2/src/main/java/com/ayakacraft/carpetayakaaddition/commands/address/Address.java
@@ -27,16 +27,21 @@ import net.minecraft.world.dimension.DimensionType;
 
 public class Address extends AbstractAddress<DimensionType> {
 
-    public Address(String id, String dim, Vec3d pos, String desc, long weight) {
+    public Address(String id, String dim, Vec3d pos, String desc, int weight) {
         super(id, dim, pos, desc, weight);
     }
 
-    public Address(String id, DimensionType dim, Vec3d pos, String desc, long weight) {
-        super(id, String.valueOf(DimensionType.getId(dim)), pos, desc, weight);
+    public Address(String id, DimensionType dim, Vec3d pos, String desc, int weight) {
+        super(id, dim, pos, desc, weight);
     }
 
     public Address(AddressOld old) {
         super(old);
+    }
+
+    @Override
+    protected String transDim(DimensionType dimensionType) {
+        return String.valueOf(DimensionType.getId(dimensionType));
     }
 
     @Override