+++ changelog.md
@@ -1,4 +1,5 @@
 ### General
 
 - Added warning for `frostWalkerNoFreezing` as enchantments are defined by datapacks in 1.21+
+- Made address weight reduce over time
 - Performance and experience improvements